Coverity scan: [fixed]: Uninitialized variables  (UNINIT) [fixed]: Error handling issues  (UNCAUGHT_EXCEPT) [fixed]: Control flow issues  (NO_EFFECT) [fixed]: Uninitialized scalar variable (UNINIT)

+++ Utilities/MBTester/mbrtutester.cc
@@ -479,7 +479,7 @@ int main( int argc, char** argv )
 
 						cout << "(reply): count=" << ModbusRTU::dat2str(ret.count) << endl;
 
-						for( int i = 0; i < ret.count; i++ )
+						for( size_t i = 0; i < ret.count; i++ )
 						{
 							cout << i << ": (" << ModbusRTU::dat2str( reg + i ) << ") = " << (int)(ret.data[i])
 								 << " ("
@@ -507,7 +507,7 @@ int main( int argc, char** argv )
 
 						cout << "(reply): count=" << ModbusRTU::dat2str(ret.count) << endl;
 
-						for( int i = 0; i < ret.count; i++ )
+						for( size_t i = 0; i < ret.count; i++ )
 						{
 							cout << i << ": (" << ModbusRTU::dat2str( reg + i ) << ") = " << (int)(ret.data[i])
 								 << " ("
